Fix native notification D-Bus crash with Qt6
QImage::bytesPerLine() returns qsizetype in Qt6, which is 64-bit on
64-bit platforms. The notification image-data structure requires a
32-bit integer for the row stride, so serializing it directly causes
D-Bus to abort on an int32/int64 type mismatch.

+++ src/widgets/osd_x11.cpp
@@ -60,7 +60,7 @@ QDBusArgument& operator<<(QDBusArgument& arg, const QImage& image) {
   arg.beginStructure();
   arg << i.width();
   arg << i.height();
-  arg << i.bytesPerLine();
+  arg << static_cast<int>(i.bytesPerLine());
   arg << i.hasAlphaChannel();
   int channels = i.isGrayscale() ? 1 : (i.hasAlphaChannel() ? 4 : 3);
   arg << i.depth() / channels;
